Added reviews for film

+++ application/models/user.go
@@ -1,7 +1,7 @@
 package models
 
 type User struct {
-	Id       int64  `json:"id" gorm:"primary_key"`
+	Id       uint   `json:"id" gorm:"primary_key"`
 	Username string `json:"username"`
 	Password string `json:"password"`
 	Email    string `json:"email"`
