+++ application/user/usecase.go
@@ -8,6 +8,6 @@ type UseCase interface {
 	Login(login string, password string) (sessionId string, err error)
 	Logout(sessionId string) error
 	Add(usr *models.User) (*models.User, error)
-	Get(id int64) (*models.User, error)
+	Get(id uint) (*models.User, error)
 	Update(user *models.User) error
 }
